refactor: rename variable names for clariy and in sync with mathematical formulas
